+++ src/main/java/com/example/house/domain/HousePicture.java
@@ -4,8 +4,8 @@ import lombok.Data;
 
 @Data
 public class HousePicture {
-    private Integer id;
-    private Integer houseId;
+    private Long id;
+    private Long houseId;
     private String cdnPrefix;
     private Integer width;
     private Integer height;
